fix(wwaudio): make the Miles stub compile on Apple targets
z_wwaudio has never compiled on any Unix target. It went unnoticed because
build-macos-zh.sh builds `--target z_generals`, which does not link libwwaudio.a,
while an iOS `cmake --build` with no target builds `all` and does. So this is a
long-standing break, not a regression.

Root cause: MilesStub.h defined HANDLE as an object-like MACRO, so CompatLib's
`typedef void *HANDLE;` expanded to `typedef void *void*;`. Everything after it -
unknown HSTREAM, CRITICAL_SECTION, missing AIL_* - followed from that first parse
failure, though a -k 0 run showed the library also genuinely lacked most of the
Miles surface.

- HANDLE becomes a typedef. An identical typedef redeclaration is legal C++, so it
  is inert whichever of CompatLib or DXVK's windows_base.h is seen first.
- The four AIL_* macros become functions: every call site is qualified
  (::AIL_release_sample_handle), and a function-like macro expands there too,
  producing `::do {} while(0)`.
- Remaining AIL_* plus the thread/event helpers are variadic function templates -
  the header cannot verify prototypes it does not have, and a template never
  conflicts with a real non-template overload.
- Stub return values are chosen so no stubbed loop can spin.
- Threads.cpp guards <process.h> to _WIN32; SoundScene.cpp casts through uintptr_t
  (value-identical on 32-bit Windows, and On_Event ignores the parameter anyway).

Windows is unaffected: the whole MilesStub.h body sits inside
`#if defined(SAGE_USE_OPENAL)`, which no Windows preset sets.

Verified by exit code, both fresh compiles: iOS `cmake --build build/ios-vulkan`
EXIT 0 after deleting the z_wwaudio objects and libwwaudio.a, and
`ninja -C build/macos-vulkan z_wwaudio g_wwaudio` EXIT 0.

+++ Core/Libraries/Source/WWVegas/WWAudio/SoundScene.cpp
@@ -199,7 +199,11 @@ SoundSceneClass::Collect_Logical_Sounds (unsigned int milliseconds, int listener
 				//	Is the sound ready to notify?
 				//
 				if (sound_obj->Allow_Notify (timestamp)) {
-					listener->On_Event (AudioCallbackClass::EVENT_LOGICAL_HEARD, (uint32)listener, (uint32)sound_obj);
+					// GeneralsX @build 27/07/2026 On_Event takes uint32 params, so route the pointers
+					// through uintptr_t first. Identical on 32-bit Windows; on 64-bit this keeps the
+					// same (already discarded - On_Event ignores param1/param2) low word without
+					// tripping -Wpointer-to-int-cast.
+					listener->On_Event (AudioCallbackClass::EVENT_LOGICAL_HEARD, (uint32)(uintptr_t)listener, (uint32)(uintptr_t)sound_obj);
 				}
 			}
 		}
